MirAL: get active window from focus controller
When setting the active window to a null window, mru_active_windows
won't be updated. This makes active_window() report that the last active
window is still active and affects things like deciding if the closing
window is active or not.

This commit makes active_window() takes the focused window from the
focus controller, which is updated when setting null window, instead.
This fixes an issue where refocus happens when the last-active window
is hidden (before being closed) despite not active anymore.