MirAL: always use active_window() to find out active window
BasicWindowManager::remove_window() was using mru_active_windows to find
out if the removing window is the active window. To be consistent with
the rest of the code, it's changed to use active_window().

This fixes the issue where refocus happens when the last-active window
is closed despite not active anymore.